pyproject.toml
[project]
name = "comfyui-lf"
description = "Custom nodes for ComfyUI with a touch of extra UX ✨ Including: history for primitives, JSON manipulation, logic switches with visual feedback, LLM chat... & more!"
version = "3.1.1"
license = {file = "LICENSE"}
dependencies = ["numpy", "openai==1.12.0", "Pillow", "torch", "requests"]

[project.urls]
Repository = "https://github.com/lucafoscili/comfyui-lf"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "lucafoscili"
DisplayName = "comfyui-lf"
Icon = ""
